Replace default hash with Blake3
Crytographic security for our hash!

Performance numbers: I profiled before/after on my laptop: 4 Mbit messages, 10
channels, 1000 clients. Looks like we're taking a 2x hit, which seems okay given
the relative time: it went from 0.2% of total to 0.4% of total running time.

I tested in parallel[1], too, and it went back to 0.26%--for a total of 0.05% .
Before we get too excited, the break-even point seems to be about 1
Mbit (according to the Blake3 docs), which means it's probably premature.

[1] https://docs.rs/blake3/0.3.0/blake3/struct.Hasher.html#method.update_with_join